Set groupDavKey before uploading card
This avoids a warning ("sending card with empty key") and fixes a problem on Kerio Connect and maybe some other CardDAV implementations:
After creating a contact in Thunderbird, syncing it back without changes and edit this contact in Thunderbird to re-upload to server ended up in a duplicated contact on server. This is caused by the fact that the vcard (which creates UID from groupDavKey) is being called before the key is set on the card, leaving CardUID null. The server in my case creates a unique CardUID for its response, but next time it is being sent, the groupDavKey is being used, resulting in another new card instead of modifying the existing one.